# openapi/public-api.yaml
openapi: 3.1.0
info:
  title: QueryTube Public API
  version: 1.0.1
  description: |
    Public API v1 is the anonymous, read-only API at the existing /api/public/... paths.
    This YAML is the source of truth for public wire paths, parameters, schemas,
    responses, and compatibility policy. The contract version is not a URL prefix;
    consumers do not need to migrate to /api/v1. Version 1.0.1 formalizes v1 and
    corrects documentation to match existing responses, without changing wire behavior.

    Owners explicitly share Public Query Sets and Public Search Runs. Search Run
    visibility is independent of Query Set publication or continued existence.
    Reads use current persisted visibility. Missing/private details return 404;
    unavailable authoritative storage returns 503, without a cached-data fallback.
    Sharing can be revoked, and a public run may still be running or have failed queries.
    429 responses indicate rate limiting; the current per-process budget is not a
    guarantee of a globally exact request allowance. Error message wording is not stable.

    Consumer boundary:
    - Run identifiers are opaque strings scoped by the owner userId; retain both.
    - Lists return { items: [...] }, ordered by startedAt descending, capped by limit.
      v1 has no cursor, next-page token, total count, or snapshot-consistency guarantee.
    - Details contain id and queryResults[].videos[], not a flat results array.
      Video Results provide videoId, url (YouTube watch URL), and title (a required
      string, possibly empty). Consumers can deduplicate by videoId across queries.
      Empty arrays are valid; order within queryResults/videos is not guaranteed.
    - All schema-defined fields are public contract, including fields beyond this
      minimum consumer subset. Optional fields may be absent or null only where
      their schema permits it. Consumers must tolerate unknown added fields.
    - Internal models, Firestore paths, Output YAML, ID encodings, and undocumented
      properties are implementation details and are not public guarantees.

    Compatibility policy:
    - Adding optional response fields, endpoints, or optional query parameters is
      normally compatible when existing requests and semantics remain unchanged.
    - Removing/renaming fields or paths, changing types/nullability incompatibly,
      making optional request fields required, making required response fields
      optional, making optional fields required in the published schema, removing
      enum values, restructuring responses, and changing endpoint/access/order/limit
      semantics are breaking. Even optional response fields cannot be removed silently.
    - Response enum additions require compatibility review because consumers may
      exhaustively handle the existing values. Do not assume they are safe additions.
    - Compatible additions increment the minor contract version; corrections that
      preserve wire behavior increment the patch version. Breaking changes require
      a new major contract and a separate OpenSpec change with migration guidance,
      consumer impact review, and an explicit deprecation/support plan. Keep v1
      available during migration; do not silently repurpose or redirect its URLs.
      A separate versioned route may be introduced when a concrete breaking change
      requires coexistence. No v2 infrastructure is required by this contract.
security: []
servers:
  - url: /
    description: Current host server
paths:
  /api/public/users/{userId}/query-sets:
    get:
      tags:
        - Query Sets
      summary: List public Query Sets
      description: Returns Public Query Set summaries for the specified user. Does not include their YAML Search Definitions.
      operationId: listPublicQuerySets
      parameters:
        - name: userId
          in: path
          required: true
          description: The Firebase UID of the Query Set owner.
          schema:
            type: string
          example: cabeceowei-uid-123
      responses:
        '200':
          description: Successfully retrieved Public Query Set list.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicQuerySetSummary'
              example:
                items:
                  - id: qs_1711892019_abc12
                    name: WoW Forever Multilingual
                    queryCount: 18
                    createdAt: '2026-10-01T01:00:00.000Z'
                    updatedAt: '2026-10-01T02:00:00.000Z'
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Authoritative storage unavailable; cached data is not returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/public/users/{userId}/query-sets/{querySetId}:
    get:
      tags:
        - Query Sets
      summary: Get public Query Set details
      description: Returns the Public Query Set, including its YAML Search Definition, if the owner has enabled Query Set publication.
      operationId: getPublicQuerySet
      parameters:
        - name: userId
          in: path
          required: true
          description: The Firebase UID of the Query Set owner.
          schema:
            type: string
          example: cabeceowei-uid-123
        - name: querySetId
          in: path
          required: true
          description: The unique identifier of the Query Set.
          schema:
            type: string
          example: qs_1711892019_abc12
      responses:
        '200':
          description: Successfully retrieved public Query Set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicQuerySet'
              example:
                id: qs_1711892019_abc12
                name: WoW Forever Multilingual
                rawYaml: "version: 1\ndefaults:\n  max_results: 10\nqueries:\n  - id: wow-review-ja\n    q: \"WoW Forever 感想\"\n    relevance_language: ja\n    region_code: JP\n"
                queryCount: 18
                createdAt: '2026-10-01T01:00:00.000Z'
                updatedAt: '2026-10-01T02:00:00.000Z'
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Query Set not found or not marked as public.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Not Found
                message: Query set not found or not public
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Authoritative storage unavailable; cached data is not returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/public/users/{userId}/search-runs:
    get:
      tags:
        - Search Runs
      summary: List public Search Runs
      description: Returns Public Search Run summaries for the specified user, ordered by startedAt descending, with no cursor or next-page token. Search Run visibility is independent of Query Set publication. Does not return input YAML or Video Result lists. If querySetId is provided, only Public Search Runs associated with that Query Set are returned; the Query Set need not still exist or be public.
      operationId: listPublicSearchRuns
      parameters:
        - name: userId
          in: path
          required: true
          description: The Firebase UID of the resource owner.
          schema:
            type: string
          example: cabeceowei-uid-123
        - name: limit
          in: query
          required: false
          description: Maximum number of Search Runs to return (1-100, default 50). Existing behavior parses the integer prefix, clamps to 1-100, and uses 50 when parsing fails or the value is empty.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          example: 50
        - name: querySetId
          in: query
          required: false
          description: Optional filter to restrict results to search runs originating from a specific Query Set.
          schema:
            type: string
          example: qs_1711892019_abc12
      responses:
        '200':
          description: Successfully retrieved list of public search runs.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicSearchRunSummary'
              example:
                items:
                  - id: run_1711893000_xyz78
                    querySetId: qs_1711892019_abc12
                    querySetName: WoW Forever Multilingual
                    status: completed
                    queryCount: 18
                    successfulQueries: 18
                    failedQueries: 0
                    totalResults: 126
                    startedAt: '2026-10-01T02:10:00.000Z'
                    completedAt: '2026-10-01T02:10:15.000Z'
                    createdAt: '2026-10-01T02:10:00.000Z'
                    visibility: public
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Authoritative storage unavailable; cached data is not returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/public/users/{userId}/search-runs/{runId}:
    get:
      tags:
        - Search Runs
      summary: Get Public Search Run details
      description: Returns the Public Search Run, including its executed YAML Search Definition, Query Results, and Video Results. The Search Run itself must have public visibility, independently of Query Set publication or continued existence.
      operationId: getPublicSearchRunDetails
      parameters:
        - name: userId
          in: path
          required: true
          description: The Firebase UID of the resource owner.
          schema:
            type: string
          example: cabeceowei-uid-123
        - name: runId
          in: path
          required: true
          description: The unique identifier of the search run.
          schema:
            type: string
          example: run_1711893000_xyz78
      responses:
        '200':
          description: Successfully retrieved public search run details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicSearchRun'
              example:
                id: run_1711893000_xyz78
                querySetId: qs_1711892019_abc12
                querySetName: WoW Forever Multilingual
                status: completed
                queryCount: 1
                successfulQueries: 1
                failedQueries: 0
                totalResults: 1
                inputYaml: "version: 1\nqueries:\n  - id: ja-review\n    q: WoW Forever 感想\n"
                startedAt: '2026-10-01T02:10:00.000Z'
                completedAt: '2026-10-01T02:10:15.000Z'
                createdAt: '2026-10-01T02:10:00.000Z'
                visibility: public
                queryResults:
                  - id: ja-review
                    sourceQueryId: ja-review
                    query: WoW Forever 感想
                    relevanceLanguage: ja
                    regionCode: JP
                    status: success
                    resultCount: 1
                    errorCode: null
                    errorMessage: null
                    startedAt: '2026-10-01T02:10:01.000Z'
                    completedAt: '2026-10-01T02:10:03.000Z'
                    videos:
                      - videoId: dQw4w9WgXcQ
                        title: WoW Forever Gameplay Review
                        channelId: UC1234567890
                        channelTitle: GameChannel
                        publishedAt: '2026-09-28T14:00:00.000Z'
                        description: An in-depth review of WoW Forever in Japanese.
                        url: https://www.youtube.com/watch?v=dQw4w9WgXcQ
                        thumbnailUrl: https://i.ytimg.com/vi/dQw4w9WgXcQ/hqdefault.jpg
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Search run not found or not marked as public.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Not Found
                message: Search run not found or not public
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Authoritative storage unavailable; cached data is not returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/public/users/{userId}/query-sets/{querySetId}/search-runs:
    get:
      tags:
        - Search Runs
      summary: List Public Search Runs associated with a Query Set (convenience endpoint)
      description: Backward-compatible convenience endpoint equivalent to GET /api/public/users/{userId}/search-runs?querySetId={querySetId}. Returns Public Search Runs associated with this Query Set, independently of Query Set publication or continued existence.
      operationId: listPublicSearchRunsByQuerySet
      parameters:
        - name: userId
          in: path
          required: true
          description: The Firebase UID of the Query Set owner.
          schema:
            type: string
          example: cabeceowei-uid-123
        - name: querySetId
          in: path
          required: true
          description: The unique identifier of the Query Set.
          schema:
            type: string
          example: qs_1711892019_abc12
        - name: limit
          in: query
          required: false
          description: Maximum number of Search Runs to return (1-100, default 50). Existing behavior parses the integer prefix, clamps to 1-100, and uses 50 when parsing fails or the value is empty.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
          example: 50
      responses:
        '200':
          description: Successfully retrieved list of search runs.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicSearchRunSummary'
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Authoritative storage unavailable; cached data is not returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PublicQuerySetSummary:
      type: object
      required:
        - id
        - name
        - queryCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier of the Query Set.
        name:
          type: string
          description: Name of the Query Set.
        queryCount:
          type: integer
          description: Number of Queries in the YAML Search Definition.
        createdAt:
          type: string
          format: date-time
          description: ISO timestamp when the Query Set was created.
        updatedAt:
          type: string
          format: date-time
          description: ISO timestamp when the Query Set was last modified.

    PublicQuerySet:
      type: object
      required:
        - id
        - name
        - rawYaml
        - queryCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier of the Query Set.
        name:
          type: string
          description: Name of the Query Set.
        rawYaml:
          type: string
          description: The saved YAML Search Definition expressed as a string.
        queryCount:
          type: integer
          description: Number of Queries in the YAML Search Definition.
        createdAt:
          type: string
          format: date-time
          description: ISO timestamp when the Query Set was created.
        updatedAt:
          type: string
          format: date-time
          description: ISO timestamp when the Query Set was last modified.

    PublicSearchRunSummary:
      type: object
      required:
        - id
        - status
        - queryCount
        - successfulQueries
        - failedQueries
        - totalResults
        - startedAt
        - createdAt
        - visibility
      properties:
        id:
          type: string
          description: Opaque Search Run identifier scoped by the owner userId; retain both when referring to a run.
        querySetId:
          type: [string, "null"]
          description: Historical association with a source Query Set, which need not still exist.
        querySetName:
          type: [string, "null"]
          description: Name of the associated Query Set at run time, which need not match its current name.
        status:
          type: string
          enum: [running, completed, partial, failed]
          description: Execution outcome status.
        queryCount:
          type: integer
          description: Number of executed queries.
        successfulQueries:
          type: integer
          description: Number of queries that completed successfully.
        failedQueries:
          type: integer
          description: Number of queries that encountered errors.
        totalResults:
          type: integer
          description: Total number of Video Results across all Query Results.
        startedAt:
          type: string
          format: date-time
          description: ISO timestamp when search execution started.
        completedAt:
          type: [string, "null"]
          format: date-time
          description: ISO timestamp when search execution finished.
        createdAt:
          type: string
          format: date-time
          description: ISO timestamp of Search Run document creation.
        visibility:
          type: string
          enum: [public]
          description: Search Run visibility; only public Search Runs are returned.
          example: public

    PublicSearchRun:
      type: object
      required:
        - id
        - status
        - queryCount
        - successfulQueries
        - failedQueries
        - totalResults
        - inputYaml
        - startedAt
        - createdAt
        - visibility
        - queryResults
      properties:
        id:
          type: string
          description: Opaque Search Run identifier scoped by the owner userId; retain both when referring to a run.
        querySetId:
          type: [string, "null"]
          description: Historical association with a source Query Set, which need not still exist.
        querySetName:
          type: [string, "null"]
          description: Name of the associated Query Set at run time, which need not match its current name.
        status:
          type: string
          enum: [running, completed, partial, failed]
          description: Execution outcome status.
        queryCount:
          type: integer
          description: Number of executed queries.
        successfulQueries:
          type: integer
          description: Number of queries that completed successfully.
        failedQueries:
          type: integer
          description: Number of queries that encountered errors.
        totalResults:
          type: integer
          description: Total number of Video Results across all Query Results.
        inputYaml:
          type: string
          description: The YAML Search Definition executed during this Search Run.
        startedAt:
          type: string
          format: date-time
          description: ISO timestamp when search execution started.
        completedAt:
          type: [string, "null"]
          format: date-time
          description: ISO timestamp when search execution finished.
        createdAt:
          type: string
          format: date-time
          description: ISO timestamp of Search Run document creation.
        visibility:
          type: string
          enum: [public]
          description: Search Run visibility; only public Search Runs are returned.
          example: public
        queryResults:
          type: array
          description: Recorded Query Results and their Video Results or failure information. May be empty; array order and video uniqueness across queries are not guaranteed.
          items:
            $ref: '#/components/schemas/QueryResult'

    QueryResult:
      type: object
      required:
        - id
        - sourceQueryId
        - query
        - status
        - resultCount
        - startedAt
        - completedAt
        - videos
      properties:
        id:
          type: string
          description: Unique identifier of the Query Result.
        sourceQueryId:
          type: string
          description: ID from the YAML query definition.
        query:
          type: string
          description: Executed query string.
        relevanceLanguage:
          type: [string, "null"]
          description: ISO 639-1 language code if specified.
        regionCode:
          type: [string, "null"]
          description: ISO 3166-1 alpha-2 country code if specified.
        status:
          type: string
          enum: [success, failed]
          description: Success or failure status of the recorded Query Outcome.
        resultCount:
          type: integer
          description: Number of Video Results in this Query Result.
        errorCode:
          type: [string, "null"]
          description: Error identifier if query failed.
        errorMessage:
          type: [string, "null"]
          description: Error details if query failed.
        startedAt:
          type: string
          anyOf:
            - const: ""
            - format: date-time
          description: ISO timestamp when this individual query began; empty string for legacy records without a timestamp.
        completedAt:
          type: string
          anyOf:
            - const: ""
            - format: date-time
          description: ISO timestamp when this individual query finished; empty string for legacy records without a timestamp.
        videos:
          type: array
          description: Video Results included in this Query Result.
          items:
            $ref: '#/components/schemas/Video'

    Video:
      type: object
      required:
        - videoId
        - title
        - channelId
        - channelTitle
        - publishedAt
        - description
        - url
        - thumbnailUrl
      properties:
        videoId:
          type: string
          description: YouTube video ID.
        title:
          type: string
          description: Recorded video title with decoded HTML entities; empty string when unavailable.
        channelId:
          type: string
          description: YouTube channel ID.
        channelTitle:
          type: string
          description: YouTube channel title.
        publishedAt:
          type: string
          anyOf:
            - const: ""
            - format: date-time
          description: Video publication timestamp; empty string when unavailable.
        description:
          type: string
          description: Video snippet description.
        url:
          type: string
          format: uri
          description: Direct link to watch the video on YouTube.
        thumbnailUrl:
          type: string
          anyOf:
            - const: ""
            - format: uri
          description: Video thumbnail image URL; empty string when unavailable.

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error classification code.
        message:
          type: string
          description: Human-readable error description.